+++ packages/vscode/src/log.ts
@@ -1,0 +1,3 @@
+import { window } from 'vscode'
+
+export const log = window.createOutputChannel('UnoCSS')
