+++ packages/vscode/src/utils.ts
@@ -1,0 +1,16 @@
+
+export function throttle<T extends((...args: any) => any)>(func: T, timeFrame: number): T {
+  let lastTime = 0
+  let timer: any
+  return function() {
+    const now = Date.now()
+    clearTimeout(timer)
+    if (now - lastTime >= timeFrame) {
+      lastTime = now
+      return func()
+    }
+    else {
+      timer = setTimeout(func, timeFrame)
+    }
+  } as T
+}
